+++ index.html
@@ -445,6 +445,8 @@
     </template>
 
 
-    <script src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
+    <script type="module" src="js/data.js"></script>
+    <script type="module" src="js/util.js"></script>
   </body>
 </html>
